% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{}
\usage{
hello(x)
}
\arguments{
\item{x}{the name of the person to say hi to}
}
\value{
the output from \code{\link{print}}
}
\description{

}
\examples{
hello("Jhon")
\dontrun{
hello("Steve")}
}
